+++ MSAL/src/native_auth/network/responses/validator/reset_password/MSALNativeAuthResetPasswordResponseValidator.swift
@@ -58,7 +58,7 @@ final class MSALNativeAuthResetPasswordResponseValidator: MSALNativeAuthResetPas
         } else if let continuationToken = response.continuationToken {
             return .success(continuationToken: continuationToken)
         } else {
-            MSALLogger.log(level: .error,
+            MSALNativeAuthLogger.log(level: .error,
                            context: context,
                            format: "resetpassword/start returned success with unexpected response body")
 
@@ -69,7 +69,7 @@ final class MSALNativeAuthResetPasswordResponseValidator: MSALNativeAuthResetPas
     private func handleStartFailed(_ error: Error,
                                    with context: MSIDRequestContext) -> MSALNativeAuthResetPasswordStartValidatedResponse {
         guard let apiError = error as? MSALNativeAuthResetPasswordStartResponseError else {
-            MSALLogger.logPII(level: .error,
+            MSALNativeAuthLogger.logPII(level: .error,
                            context: context,
                            format: "resetpassword/start: Unable to decode error response: \(MSALLogMask.maskPII(error))")
 
@@ -128,20 +128,20 @@ final class MSALNativeAuthResetPasswordResponseValidator: MSALNativeAuthResetPas
                     continuationToken
                 )
             } else {
-                MSALLogger.log(level: .error, context: context, format: "Missing expected fields from backend")
+                MSALNativeAuthLogger.log(level: .error, context: context, format: "Missing expected fields from backend")
                 return .unexpectedError(.init(errorDescription: MSALNativeAuthErrorMessage.unexpectedResponseBody))
             }
         case .password,
              .otp:
             let errorDescription = MSALNativeAuthErrorMessage.unexpectedChallengeType
-            MSALLogger.log(level: .error, context: context, format: errorDescription)
+            MSALNativeAuthLogger.log(level: .error, context: context, format: errorDescription)
             return .unexpectedError(.init(errorDescription: errorDescription))
         }
     }
 
     private func handleChallengeError(_ error: Error, with context: MSIDRequestContext) -> MSALNativeAuthResetPasswordChallengeValidatedResponse {
         guard let apiError = error as? MSALNativeAuthResetPasswordChallengeResponseError else {
-            MSALLogger.logPII(
+            MSALNativeAuthLogger.logPII(
                 level: .error,
                 context: context,
                 format: "resetpassword/challenge: Unable to decode error response: \(MSALLogMask.maskPII(error))"
@@ -176,7 +176,7 @@ final class MSALNativeAuthResetPasswordResponseValidator: MSALNativeAuthResetPas
 
     private func handleContinueError(_ error: Error, with context: MSIDRequestContext) -> MSALNativeAuthResetPasswordContinueValidatedResponse {
         guard let apiError = error as? MSALNativeAuthResetPasswordContinueResponseError else {
-            MSALLogger.logPII(
+            MSALNativeAuthLogger.logPII(
                 level: .error,
                 context: context,
                 format: "resetpassword/continue: Unable to decode error response: \(MSALLogMask.maskPII(error))"
@@ -192,7 +192,7 @@ final class MSALNativeAuthResetPasswordResponseValidator: MSALNativeAuthResetPas
              .invalidRequest:
             return .error(apiError)
         case .verificationRequired:
-            MSALLogger.log(level: .error, context: context, format: "verificationRequired is not supported yet")
+            MSALNativeAuthLogger.log(level: .error, context: context, format: "verificationRequired is not supported yet")
             return .unexpectedError(nil)
         case .unknown:
             return .unexpectedError(apiError)
@@ -221,7 +221,7 @@ final class MSALNativeAuthResetPasswordResponseValidator: MSALNativeAuthResetPas
 
     private func handleSubmitError(_ error: Error, with context: MSIDRequestContext) -> MSALNativeAuthResetPasswordSubmitValidatedResponse {
         guard let apiError = error as? MSALNativeAuthResetPasswordSubmitResponseError else {
-            MSALLogger.logPII(
+            MSALNativeAuthLogger.logPII(
                 level: .error,
                 context: context,
                 format: "resetpassword/submit: Unable to decode error response: \(MSALLogMask.maskPII(error))"
@@ -271,7 +271,7 @@ final class MSALNativeAuthResetPasswordResponseValidator: MSALNativeAuthResetPas
         with context: MSIDRequestContext
     ) -> MSALNativeAuthResetPasswordPollCompletionValidatedResponse {
         guard let apiError = error as? MSALNativeAuthResetPasswordPollCompletionResponseError else {
-            MSALLogger.logPII(
+            MSALNativeAuthLogger.logPII(
                 level: .error,
                 context: context,
                 format: "resetpassword/poll_completion: Unable to decode error response: \(MSALLogMask.maskPII(error))"