+++ MSAL/src/native_auth/network/responses/validator/token/MSALNativeAuthTokenResponseValidator.swift
@@ -59,7 +59,7 @@ final class MSALNativeAuthTokenResponseValidator: MSALNativeAuthTokenResponseVal
         case .failure(let tokenResponseError):
             guard let tokenResponseError =
                     tokenResponseError as? MSALNativeAuthTokenResponseError else {
-                MSALLogger.logPII(
+                MSALNativeAuthLogger.logPII(
                     level: .error,
                     context: context,
                     format: "Token: Unable to decode error response: \(MSALLogMask.maskPII(tokenResponseError))")
@@ -102,7 +102,7 @@ final class MSALNativeAuthTokenResponseValidator: MSALNativeAuthTokenResponseVal
                     return .error(.invalidOOBCode(responseError))
                 } else if responseError.subError == .mfaRequired {
                     guard let continuationToken = responseError.continuationToken else {
-                        MSALLogger.log(
+                        MSALNativeAuthLogger.log(
                             level: .error,
                             context: context,
                             format: "Token: MFA required response, expected continuation token not empty")
@@ -113,7 +113,7 @@ final class MSALNativeAuthTokenResponseValidator: MSALNativeAuthTokenResponseVal
                     return .strongAuthRequired(continuationToken: continuationToken)
                 } else if responseError.subError == .jitRequired {
                     guard let continuationToken = responseError.continuationToken else {
-                        MSALLogger.log(
+                        MSALNativeAuthLogger.log(
                             level: .error,
                             context: context,
                             format: "Token: JIT required response, expected continuation token not empty")
@@ -183,7 +183,7 @@ final class MSALNativeAuthTokenResponseValidator: MSALNativeAuthTokenResponseVal
         errorCodesConverterFunction: (MSALNativeAuthESTSApiErrorCodes, MSALNativeAuthTokenResponseError) -> MSALNativeAuthTokenValidatedErrorType
     ) -> MSALNativeAuthTokenValidatedResponse {
         guard var errorCodes = apiError.errorCodes, !errorCodes.isEmpty else {
-            MSALLogger.log(level: .error, context: context, format: "/token error - Empty error_codes received")
+            MSALNativeAuthLogger.log(level: .error, context: context, format: "/token error - Empty error_codes received")
             return useInvalidRequestAsDefaultResult ? .error(.invalidRequest(apiError)) : .error(.generalError(apiError))
         }
 
@@ -193,7 +193,7 @@ final class MSALNativeAuthTokenResponseValidator: MSALNativeAuthTokenResponseVal
         if let knownErrorCode = MSALNativeAuthESTSApiErrorCodes(rawValue: firstErrorCode) {
             validatedResponse = .error(errorCodesConverterFunction(knownErrorCode, apiError))
         } else {
-            MSALLogger.logPII(
+            MSALNativeAuthLogger.logPII(
                 level: .error,
                 context: context,
                 format: "/token error - Unknown code received in error_codes: \(MSALLogMask.maskPII(firstErrorCode))"
@@ -212,7 +212,7 @@ final class MSALNativeAuthTokenResponseValidator: MSALNativeAuthTokenResponseVal
                 errorMessage = "/token error - Unknown ESTS received in error_codes with code: \(MSALLogMask.maskPII(errorCode)) (ignoring)"
             }
 
-            MSALLogger.logPII(level: .verbose, context: context, format: errorMessage)
+            MSALNativeAuthLogger.logPII(level: .verbose, context: context, format: errorMessage)
         }
 
         return validatedResponse