+++ MSAL/src/native_auth/network/responses/validator/token/validated_response/MSALNativeAuthTokenValidatedResponse.swift
@@ -90,7 +90,7 @@ enum MSALNativeAuthTokenValidatedErrorType: Error {
                 errorUri: apiError.errorURI
             )
         case .expiredRefreshToken(let apiError):
-            MSALLogger.logPII(level: .error, context: nil, format: "Error not treated - \(MSALLogMask.maskPII(self))")
+            MSALNativeAuthLogger.logPII(level: .error, context: nil, format: "Error not treated - \(MSALLogMask.maskPII(self))")
             return SignInStartError(
                 type: .generalError,
                 message: apiError.errorDescription,
@@ -138,7 +138,7 @@ enum MSALNativeAuthTokenValidatedErrorType: Error {
         case .userNotFound(let apiError),
              .invalidPassword(let apiError),
              .invalidOOBCode(let apiError):
-            MSALLogger.logPII(level: .error, context: nil, format: "Error not treated - \(MSALLogMask.maskPII(self))")
+            MSALNativeAuthLogger.logPII(level: .error, context: nil, format: "Error not treated - \(MSALLogMask.maskPII(self))")
             return RetrieveAccessTokenError(
                 type: .generalError,
                 message: apiError.errorDescription,