+++ MSAL/src/public/configuration/global/MSALLoggerConfig.h
@@ -56,10 +56,6 @@ NS_ASSUME_NONNULL_BEGIN
  The minimum log level for messages to be passed onto the log callback.
  */
 @property (atomic) MSALLogLevel logLevel;
-/**
- MSAL provides logging callbacks that assist in diagnostics. There is a boolean value in the logging callback that indicates whether the message contains user information. If piiEnabled is set to NO, the callback will not be triggered for log messages that contain any user information. By default the library will not return any messages with user information in them.
- */
-@property (nonatomic) BOOL piiEnabled DEPRECATED_MSG_ATTRIBUTE("Use logMaskingLevel instead");
 
 /**
  MSAL provides logging callbacks that assist in diagnostics. By default the library will not return any messages with any user or organizational information. However, this might make diagnosing issues difficult.